feat!: reverse roles of headline and section
Headline used to be a GreaterElement with Section nested inside. This
pushed title into its own field in the Headline and that caused
inconveniences because unist traversing utilities only visit children
field, so they missed the title.

BREAKING CHANGE: The roles of Headline and Section are now
reversed. Section now contains Headline (always as the first
children), Planning (as a possible second children),
PropertyDrawer (as a possible third children), and then the rest of
the elements in the section.

BREAKING CHANGE: Headline now does not have a title field and title is
contents is now placed inside children field.

BREAKING CHANGE: orgast-util-visit-ids now passes Section instead of
Headline as the second argument of callback.

This should hopefully make it more convenient to implement features
that traverse all of the tree (smartypants, or visiting all
links). They would likely miss titles before, but not now.

If you are using uniorg-parse, uniorg-extract-keywords, uniorg-slug,
uniorg-rehype, nothing changes for you. However, if you use
orgast-util-visit-ids or manipulate the tree directly, the changes
above are breaking.