fix(uniorg-parse): fix parsing lists nested in quotes nested in lists
When parsing the test case as follows

    - list 1
      #+begin_quote
      - list 2
      #+end_quote

the parser threw an exception because "list 2" item couldn't find
itself in the structure that was passed in. The bug was introduced
because uniorg now properly skips blocks within list items, so "list
2" item is not present in the structure of "list 1".

After a second examination of org-element.el, it is only passing
structure if element is list or item.