Adding className if heading is a todo
This conditionally adds a className of "todo-item" to headings which contain a TODO face. This will allow for better styling controls, as currently there isn't a way to selectively style headings that are TODO items separately from headings that are not TODO items. This should result in a familiar output compared to org-export, which does mark headings with TODO items with an additional class.